# Conflicts:
#	lib/src/main/java/fr/onema/lib/sensor/position/Pressure.java
#	lib/src/main/java/fr/onema/lib/virtualizer/entry/ReferenceEntry.java
#	lib/src/main/java/fr/onema/lib/virtualizer/entry/VirtualizerEntry.java

+++ lib/src/main/java/fr/onema/lib/geo/CartesianCoordinate.java
@@ -5,12 +5,18 @@ package fr.onema.lib.geo;
  */
 // TODO : complete
 public class CartesianCoordinate {
-    public final int x;
-    public final int y;
-    public final int z;
+    public final double x;
+    public final double y;
+    public final double z;
 
-    // TODO : complete
-    public CartesianCoordinate(int x, int y, int z) {
+    /**
+     * Construit une coordonnée cartésienne
+     *
+     * @param x la valeur sur x
+     * @param y la valeur sur y
+     * @param z la valeur sur z
+     */
+    public CartesianCoordinate(double x, double y, double z) {
         this.x = x;
         this.y = y;
         this.z = z;
